services/video_service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
httpx
python-jose[cryptography]
